requirements.txt
Django==5.2.7
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.0.0
